refactor(checker)!: rename ac_guard.checker -> ac_guard.code_gate
The module's role is the git hook-time code quality gate (pass/fail
orchestration in pre-commit / pre-push). `checker` conflates with the
runtime enforcer and with generic "type checker" noun; `code_gate`
names the responsibility directly: code domain + binary gate.

Scope:
- Rename src/ac_guard/checker/ -> src/ac_guard/code_gate/
- Rename tests/unit/test_checker/ -> tests/unit/test_code_gate/
- Rename tests/integration/test_checker_e2e.py
- Update all ac_guard.checker imports across src/ and tests/
- Update .importlinter layers + config-encapsulation contracts
- Update module/role docstrings

No backward-compat shim. Enforcer rename follows in the next commit.
