+++ src/ac_guard/cli/status.py
@@ -17,7 +17,7 @@ from typing import TYPE_CHECKING
 
 from ac_guard import __version__
 from ac_guard.adapters.registry import get_adapter, list_adapters
-from ac_guard.checker.core import BUCKET_AWARE_STAGES, detect_language
+from ac_guard.code_gate.core import BUCKET_AWARE_STAGES, detect_language
 from ac_guard.config import ConfigError, load_config, resolve_config
 from ac_guard.generator.core import read_state
