+++ tests/integration/test_code_gate_e2e.py
@@ -1,4 +1,4 @@
-"""Integration tests for Phase 3 Checker end-to-end (WP3.4).
+"""Integration tests for Phase 3 code_gate end-to-end (WP3.4).
 
 Test matrix dimensions:
 - CLI commands: check / verify / run / gate
@@ -18,8 +18,8 @@ import pytest
 import yaml
 from typer.testing import CliRunner
 
-from ac_guard.checker.models import CheckResult, StageOutcome
 from ac_guard.cli.main import app
+from ac_guard.code_gate.models import CheckResult, StageOutcome
 from ac_guard.reporter.formatting import format_markdown
 
 runner = CliRunner()
@@ -91,18 +91,18 @@ class TestFullCheckLifecycle:
         assert r.exit_code == 0
 
         # check (commit stage)
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["check", "-c", str(config)])
         assert r.exit_code == 0
         assert "PASSED" in r.output
 
         # verify (push stage)
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["verify", "-c", str(config)])
         assert r.exit_code == 0
 
         # gate run (commit)
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(
                 app, ["gate", "run", "-s", "pre-commit", "-c", str(config)]
             )
@@ -121,7 +121,7 @@ class TestFullCheckLifecycle:
         (tmp_path / ".git").mkdir()
         runner.invoke(app, ["install", "-a", "claude-code", "-c", str(config)])
 
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["check", "-c", str(config)])
         assert r.exit_code == 0
 
@@ -129,7 +129,7 @@ class TestFullCheckLifecycle:
         _config_with_checks(tmp_path, fail=True)
         runner.invoke(app, ["update", "-c", str(config)])
 
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["check", "-c", str(config)])
         assert r.exit_code == 1
         assert "FAILED" in r.output
@@ -146,7 +146,7 @@ class TestCheckCommand:
     def test_default_config_passes(self, tmp_path: Path) -> None:
         """B1: Default config with no files → all checks pass."""
         config = _write_config(tmp_path)
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["check", "-c", str(config)])
         assert r.exit_code == 0
         assert "PASSED" in r.output
@@ -154,7 +154,7 @@ class TestCheckCommand:
     def test_custom_check_fails(self, tmp_path: Path) -> None:
         """B2: Failing custom check → exit 1 + FAILED."""
         config = _config_with_checks(tmp_path, fail=True)
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["check", "-c", str(config)])
         assert r.exit_code == 1
         assert "FAILED" in r.output
@@ -163,7 +163,7 @@ class TestCheckCommand:
     def test_explicit_files(self, tmp_path: Path) -> None:
         """B3: --files passes explicit file list."""
         config = _write_config(tmp_path)
-        with patch("ac_guard.checker.core.shutil.which", return_value=None):
+        with patch("ac_guard.code_gate.core.shutil.which", return_value=None):
             r = runner.invoke(app, ["check", "--files", "a.py", "-c", str(config)])
         assert r.exit_code == 0
 
@@ -184,14 +184,14 @@ class TestVerifyCommand:
     def test_push_all_pass(self, tmp_path: Path) -> None:
         """C1: Push stage all passing → exit 0."""
         config = _config_with_checks(tmp_path, fail=False)
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["verify", "-c", str(config)])
         assert r.exit_code == 0
 
     def test_commit_fail_fast(self, tmp_path: Path) -> None:
         """C2: Commit failure → push not reached."""
         config = _config_with_checks(tmp_path, fail=True)
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["verify", "-c", str(config)])
         assert r.exit_code == 1
         # Push-only check should NOT appear (fail-fast)
@@ -200,7 +200,7 @@ class TestVerifyCommand:
     def test_skip_build(self, tmp_path: Path) -> None:
         """C3: --skip-build skips build step."""
         config = _write_config(tmp_path, {"build": {"command": "exit 1"}})
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["verify", "--skip-build", "-c", str(config)])
         assert r.exit_code == 0
 
@@ -216,7 +216,7 @@ class TestVerifyCommand:
                 },
             },
         )
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["verify", "-c", str(config)])
         assert r.exit_code == 0
         assert "build" in r.output.lower()
@@ -248,7 +248,7 @@ class TestVerifyCommand:
                 },
             },
         )
-        with patch("ac_guard.checker.core.get_changed_files", return_value=["a.py"]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=["a.py"]):
             r = runner.invoke(app, ["verify", "-c", str(config), "--format", "json"])
         assert r.exit_code == 1
         data = json_mod.loads(r.output)
@@ -271,14 +271,14 @@ class TestRunCommand:
     def test_builtin_format(self, tmp_path: Path) -> None:
         """D1: Running built-in format (skipped with no files)."""
         config = _write_config(tmp_path)
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["run", "format", "-c", str(config)])
         assert r.exit_code == 0
 
     def test_custom_check_pass(self, tmp_path: Path) -> None:
         """D2: Running a passing custom check."""
         config = _config_with_checks(tmp_path, fail=False)
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["run", "custom-commit", "-c", str(config)])
         assert r.exit_code == 0
         assert "custom-commit" in r.output
@@ -286,7 +286,7 @@ class TestRunCommand:
     def test_custom_check_fail(self, tmp_path: Path) -> None:
         """D3: Running a failing custom check."""
         config = _config_with_checks(tmp_path, fail=True)
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["run", "custom-commit", "-c", str(config)])
         assert r.exit_code == 1
 
@@ -309,7 +309,7 @@ class TestGateCommand:
     def test_commit_passed(self, tmp_path: Path) -> None:
         """E1: Commit stage pass → minimal 'passed' + exit 0."""
         config = _write_config(tmp_path)
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(
                 app, ["gate", "run", "-s", "pre-commit", "-c", str(config)]
             )
@@ -321,7 +321,7 @@ class TestGateCommand:
     def test_commit_failed(self, tmp_path: Path) -> None:
         """E2: Commit stage fail → minimal 'FAILED' + exit 1."""
         config = _config_with_checks(tmp_path, fail=True)
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(
                 app, ["gate", "run", "-s", "pre-commit", "-c", str(config)]
             )
@@ -331,7 +331,7 @@ class TestGateCommand:
     def test_push_passed(self, tmp_path: Path) -> None:
         """E3: Push stage pass → exit 0."""
         config = _config_with_checks(tmp_path, fail=False)
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["gate", "run", "-s", "pre-push", "-c", str(config)])
         assert r.exit_code == 0
 
@@ -367,11 +367,11 @@ class TestMultiLanguageE2E:
             ),
             encoding="utf-8",
         )
-        with patch("ac_guard.checker.core.run_precommit") as mock_run:
+        with patch("ac_guard.code_gate.core.run_precommit") as mock_run:
             stub = CheckResult(name="stub", passed=True, duration_ms=0)
             mock_run.return_value = stub
             with patch(
-                "ac_guard.checker.core.get_changed_files", return_value=["a.py"]
+                "ac_guard.code_gate.core.get_changed_files", return_value=["a.py"]
             ):
                 r = runner.invoke(app, ["verify", "-c", str(config)])
         assert r.exit_code == 0
@@ -386,10 +386,10 @@ class TestMultiLanguageE2E:
     def test_single_language_auto_populated(self, tmp_path: Path) -> None:
         """Only project.language configured → auto-populate fills defaults."""
         config = _write_config(tmp_path)  # project.language: python, no languages
-        with patch("ac_guard.checker.core.run_precommit") as mock_run:
+        with patch("ac_guard.code_gate.core.run_precommit") as mock_run:
             mock_run.return_value = CheckResult(name="stub", passed=True, duration_ms=0)
             with patch(
-                "ac_guard.checker.core.get_changed_files", return_value=["a.py"]
+                "ac_guard.code_gate.core.get_changed_files", return_value=["a.py"]
             ):
                 r = runner.invoke(app, ["check", "-c", str(config)])
         assert r.exit_code == 0
@@ -505,7 +505,7 @@ class TestLocalePropagation:
                 },
             },
         )
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["check", "-c", str(config)])
         assert r.exit_code == 0
         assert "阶段: pre-commit — 通过" in r.output
@@ -514,7 +514,7 @@ class TestLocalePropagation:
 
     def test_default_locale_keeps_english_output(self, tmp_path: Path) -> None:
         config = _write_config(tmp_path)  # no locale → defaults to en
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["check", "-c", str(config)])
         assert r.exit_code == 0
         assert "PASSED" in r.output
@@ -528,7 +528,7 @@ class TestReportFormats:
     def test_terminal_has_details(self, tmp_path: Path) -> None:
         """F1: Terminal output includes check details and counts."""
         config = _config_with_checks(tmp_path, fail=True)
-        with patch("ac_guard.checker.core.get_changed_files", return_value=[]):
+        with patch("ac_guard.code_gate.core.get_changed_files", return_value=[]):
             r = runner.invoke(app, ["check", "-c", str(config)])
         assert "FAIL" in r.output
         assert "custom-commit" in r.output
